+++ src/realm/object-store/impl/realm_coordinator.cpp
@@ -108,8 +108,11 @@ void RealmCoordinator::create_sync_session()
                 REALM_ASSERT(path != m_config.path);
                 REALM_ASSERT(m_config.sync_config);
                 auto copy_config = m_config;
+                copy_config.schema = util::none;
+                copy_config.realm_data = BinaryData{};
+                copy_config.audit_factory = {};
                 copy_config.path = path;
-                // Do not use seamless loss mode on the fresh Realm. Use manual mode so that
+                // Do not use 'discard local' mode on the fresh Realm. Use manual mode so that
                 // any error during the download is propagated back to the original session.
                 // This prevents a cycle if the fresh copy itself experiences a client reset.
                 copy_config.sync_config->client_resync_mode = ClientResyncMode::Manual;